requirements.txt
tensorflow-cpu>=2.10.0
streamlit>=1.20.0
numpy>=1.23.0
matplotlib>=3.6.0
seaborn>=0.12.0
scikit-learn>=1.2.0
Pillow>=9.4.0
